+++ src/config.js
@@ -22,65 +22,45 @@ module.exports = {
   ],
 
   /**
-    Max energy available for Creeps Tier 1.
-    capped to 800
+    Tiers cap and threshold
     */
   tier1_energy_cap: function() {
     return 800;
   },
-
-  /**
-    Max energy available for Creeps Tier 2.
-    capped to 1500
-    */
-  tier2_energy_cap: function() {
-    return 1500;
-  },
-
-  /**
-    Creeps Tier 2 allowed only at 1000 energy max capacity for current room.
-    */
   tier2_energy_threshold: function() {
     return 800;
   },
-
-  /**
-    Creeps Tier 3 allowed only at 1500 energy max capacity for current room.
-    */
+  tier2_energy_cap: function() {
+    return 1500;
+  },
   tier3_energy_threshold: function() {
     return 1500;
   },
 
   /**
-    Harvesters are spawn in presence of active energy sources.
-    Range [0, 8]
+    Units cap, calculated on necessity
     @param room ROOM
     @param creepsCount Object, key/value count of creeps in the current room
     */
-  harvesters_cap: function(room, creepsCount) {
+  harvesters_cap: function(room, creepsCount) { // Range [0, 8]
     const haulers = creepsCount[HAULER] ? creepsCount[HAULER] : 0;
     const miners = creepsCount[MINER] ? creepsCount[MINER] : 0;
     const tier2 =  haulers + miners;
     const sources = room.find(FIND_SOURCES_ACTIVE);
     return 4 * sources.length - tier2;
   },
-
-  /**
-    Builders are spawn only if required.
-    Range [0, 4]
-    */
-  builders_cap: function(room) {
-    // const maxEnergy = room.energyCapacityAvailable;
+  builders_cap: function(room) { // Range [0, 4]
     const constructionSites = room.find(FIND_CONSTRUCTION_SITES);
     return _.min([constructionSites.length, 4]);
   },
-
-  /**
-    Upgraders are reduced in number if builders are required
-    Range [2, 4]
-    */
-  upgraders_cap: function(room) {
+  upgraders_cap: function(room) { // Range [2, 4]
     return _.max([2, 4 - this.builders_cap(room)]);
+  },
+  haulers_cap: function(room) {
+    return this.miners_cap(room) * 2; // [0, Number of Miners * 2]
+  },
+  miners_cap: function(room) {
+    return room.find(FIND_SOURCES).length; // [0, Number of Energy Sources]
   },
 
   /**